Move wevt to its own submodule
