+++ dissect/eventlog/wevt/__init__.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from dissect.eventlog.wevt.wevt import CRIM, MAPS_WEVT_TYPE, TTBL_WEVT_TYPE, WEVT, WEVT_TYPE
+from dissect.eventlog.wevt.wevt_object import WevtObject
+
+__all__ = (
+    "CRIM",
+    "MAPS_WEVT_TYPE",
+    "TTBL_WEVT_TYPE",
+    "WEVT",
+    "WEVT_TYPE",
+    "WevtObject",
+)
