+++ dissect/eventlog/wevt/wevt_object.py
