+++ tests/_utils.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from dissect.cstruct import cstruct
 
-from dissect.eventlog.wevt import header_def
-from dissect.eventlog.wevt_object import wevt_object_def
+from dissect.eventlog.wevt.wevt import header_def
+from dissect.eventlog.wevt.wevt_object import wevt_object_def
 
 definitions = cstruct().load(header_def + wevt_object_def)
