+++ tests/wevt/test_wevtobj.py
@@ -4,9 +4,8 @@ from unittest.mock import call, patch
 
 import pytest
 
-from dissect.eventlog import wevt_object
-
-from ._utils import TEMP_HEADER, create_data_item, create_header_type
+from dissect.eventlog.wevt import wevt_object
+from tests._utils import TEMP_HEADER, create_data_item, create_header_type
 
 signatures = ["CHAN", "OPCO", "LEVL", "KEYW"]
